elizabeth/core/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
aiomysql==0.2.0
redis==5.1.1
openai==1.51.0
python-dotenv==1.0.1
websockets==13.1
pydantic==2.9.2
httpx==0.27.2
python-multipart==0.0.12
telnetlib3==2.0.4
